pyproject.toml
[build-system]
requires = ["setuptools>=62.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "datafast"
version = "0.0.1"
description = "A Python package for synthetic text dataset generation"
readme = "README.md"
license = { file = "LICENSE" }
authors = [
  { name = "Patrick Fleith" }
]
keywords = ["dataset", "text", "generation"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: GNU Affero General Public License v3",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
]
dependencies = [
    "datasets",
    "instructor",
    "google-generativeai",
    "python-dotenv",
    "anthropic",
    "openai",
    "pydantic"
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "ruff>=0.9.0",
]

[tool.setuptools.packages.find]
include = ["datafast*"]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"

[tool.ruff]
# Basic Ruff configuration
exclude = ["venv", ".venv", "build", "dist", "__pycache__"]
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "C", "W"]
ignore = ["E501"]


[project.urls]
"Documentation" = "https://github.com/patrickfleith/datafast"
"Issue Tracker" = "https://github.com/patrickfleith/datafast/issues"
"Discussions" = "https://github.com/patrickfleith/datafast/discussions"
